fix(parser): prod build fatal error in client due to undesired code optimizations
