+++ libs/parser/src/parser/parser.ts
@@ -4,17 +4,9 @@
  *  - https://golang.org/ref/spec
  *  - https://github.com/antlr/grammars-v4/tree/master/golang
  */
-import {
-  IToken,
-  TokenType,
-  Lexer,
-  Parser,
-  tokenMatcher,
-  createToken,
-  IMultiModeLexerDefinition,
-} from 'chevrotain';
-import { Identifier } from '../ast/expr-nodes';
-import { UnicodeCategory } from './unicode';
+import { Parser, tokenMatcher } from 'chevrotain';
+import * as Token from './tokens';
+import { lexerDefinition } from './lexer-definition';
 
 /////////////////////////////
 //        TOKENS           //
@@ -27,387 +19,12 @@ import { UnicodeCategory } from './unicode';
 // const Token = require('./tokens');
 
 // https://golang.org/ref/spec#Identifiers
-// identifier = letter { letter | unicode_digit }
-const IdentifierName = createToken({
-  name: 'IdentifierName',
-  label: 'Identifier',
-  // TODO: should we change this regex for manual parsing to avoid perf issues?
-  pattern: new RegExp(
-    `[_${UnicodeCategory.Letter}][_${UnicodeCategory.Letter}${
-      UnicodeCategory.DecimalDigit
-    }]*`
-  ),
-});
-
-const True = createToken({
-  name: 'True',
-  label: 'true',
-  pattern: /true/,
-});
-
-const False = createToken({
-  name: 'False',
-  label: 'false',
-  pattern: /false/,
-});
-
-const Nullable = createToken({
-  name: 'Nullable',
-  label: 'Nullable',
-  pattern: Lexer.NA,
-});
-
-const Null = createToken({
-  name: 'Null',
-  label: 'null',
-  longer_alt: IdentifierName,
-  categories: Nullable,
-  pattern: /null/,
-});
-
-const Nil = createToken({
-  name: 'Nil',
-  label: 'nil',
-  longer_alt: IdentifierName,
-  categories: Nullable,
-  pattern: /nil/,
-});
-
-const LCurly = createToken({
-  name: 'LCurly',
-  label: '{',
-  pattern: /{/,
-});
-
-const RCurly = createToken({
-  name: 'RCurly',
-  label: '}',
-  pattern: /}/,
-});
-
-const LParen = createToken({
-  name: 'LParen',
-  label: '(',
-  pattern: /\(/,
-});
-
-const RParen = createToken({
-  name: 'RParen',
-  label: ')',
-  pattern: /\)/,
-});
-
-const LSquare = createToken({
-  name: 'LSquare',
-  label: '[',
-  pattern: /\[/,
-});
-
-const RSquare = createToken({
-  name: 'RSquare',
-  label: ']',
-  pattern: /]/,
-});
-
-const Dot = createToken({
-  name: 'Dot',
-  label: '.',
-  pattern: /\./,
-});
-
-const Comma = createToken({
-  name: 'Comma',
-  label: ',',
-  pattern: /,/,
-});
-
-const TernaryOper = createToken({
-  name: 'TernaryOper',
-  label: '?',
-  pattern: /\?/,
-});
-
-const Colon = createToken({
-  name: 'Colon',
-  label: ':',
-  pattern: /:/,
-});
-
-const StringLiteral = createToken({
-  name: 'StringLiteral',
-  label: 'StringLiteral',
-  pattern: Lexer.NA,
-});
-
-const DblQuoteStringLiteral = createToken({
-  name: 'DblQuoteStringLiteral',
-  label: 'StringLiteral',
-  categories: StringLiteral,
-  pattern: /"(?:[^\\"]|\\(?:[bfnrtv"\\/]|u[0-9a-fA-F]{4}))*"/,
-});
-
-const SingleQuoteStringLiteral = createToken({
-  name: 'SingleQuoteStringLiteral',
-  label: 'StringLiteral',
-  categories: StringLiteral,
-  pattern: /'(?:[^\\']|\\(?:[bfnrtv'\\/]|u[0-9a-fA-F]{4}))*'/,
-});
-
-const NestedDblQuoteStringLiteral = createToken({
-  name: 'NestedDblQuoteStringLiteral',
-  label: 'StringLiteral',
-  categories: StringLiteral,
-  pattern: /\\"(?:[^\\\\"]|\\\\(?:[bfnrtv"\\/]|u[0-9a-fA-F]{4}))*\\"/,
-});
-
-function matchStringTemplateOpen(text: string, startOffset?: number, tokens?: IToken[]) {
-  if (tokens.length <= 2) {
-    return null;
-  }
-  const isLexingJson = tokens.find((token, index, tokenArr) => {
-    const prevToken = index - 1 >= 0 ? tokenArr[index - 1] : null;
-    return (
-      prevToken && tokenMatcher(token, Colon) && tokenMatcher(prevToken, StringLiteral)
-    );
-  });
-  if (!isLexingJson) {
-    return null;
-  }
-  return /^"{{/.exec(text.substr(startOffset));
-}
-
-const StringTemplateOpen = createToken({
-  name: 'StringTemplateOpen',
-  label: 'Open String template ("{{)',
-  pattern: { exec: matchStringTemplateOpen },
-  line_breaks: false,
-  push_mode: 'string_template',
-});
-
-const StringTemplateClose = createToken({
-  name: 'StringTemplateClose',
-  label: 'Closing StringTemplate (}}")',
-  pattern: /}}"/,
-  pop_mode: true,
-});
-
-const InlineObjectExprOpen = createToken({
-  name: 'InlineObjectExprOpen',
-  label: 'Open inline object expression',
-  pattern: /"=/,
-  push_mode: 'inline_object_expr',
-});
-
-const InlineObjectExprClose = createToken({
-  name: 'InlineObjectExprClose',
-  label: 'Close inline object expression',
-  pattern: /"/,
-  pop_mode: true,
-});
-
-const NumberLiteral = createToken({
-  name: 'NumberLiteral',
-  label: 'NumberLiteral',
-  pattern: /-?(0|[1-9]\d*)(\.\d+)?([eE][+-]?\d+)?/,
-});
-
-const WhiteSpace = createToken({
-  name: 'WhiteSpace',
-  label: 'Whitespace',
-  pattern: /\s+/,
-  group: Lexer.SKIPPED,
-  line_breaks: true,
-});
-
-const Lookup = createToken({
-  name: 'Lookup',
-  label: '$',
-  pattern: /\$/,
-});
-
-const RESOLVER_PATTERN = new RegExp(
-  `[_${UnicodeCategory.Letter}][_\.${UnicodeCategory.Letter}${
-    UnicodeCategory.DecimalDigit
-  }]*`
-);
-function matchResolverIdentifier(text: string, startOffset?: number, tokens?: IToken[]) {
-  if (tokens.length < 3) {
-    return null;
-  }
-  const lastTokenIdx = tokens.length - 1;
-  const isLexingResolver =
-    tokenMatcher(tokens[lastTokenIdx], LSquare) &&
-    tokenMatcher(tokens[lastTokenIdx - 1], IdentifierName) &&
-    tokenMatcher(tokens[lastTokenIdx - 2], Lookup);
-  if (isLexingResolver) {
-    return RESOLVER_PATTERN.exec(text.substr(startOffset));
-  }
-  return null;
-}
-
-const ResolverIdentifier = createToken({
-  name: 'ResolverIdentifier',
-  label: 'ResolverIdentifier',
-  pattern: { exec: matchResolverIdentifier },
-  line_breaks: false,
-});
-
-const BinaryOp = createToken({
-  name: 'BinaryOp',
-  label: 'Binary operator',
-  pattern: Lexer.NA,
-});
 
 // TODO: are all operators supported?
-// OPERATOR PRECEDENCE: 5 (greatest)
-const MulOp = createToken({
-  name: 'MulOp',
-  pattern: /\*|\/|%|<<|>>|&/,
-  categories: BinaryOp,
-});
 
 // OPERATOR PRECEDENCE: 4
-// TODO: are all operators supported?
-const AddOp = createToken({
-  name: 'AddOp',
-  pattern: /[+\-]/,
-  categories: BinaryOp,
-});
 
 // OPERATOR PRECEDENCE: 3
-// TODO: are all operators supported?
-const RelOp = createToken({
-  name: 'RelOp',
-  pattern: /==|!=|<=|>=|<|>/,
-  categories: BinaryOp,
-});
-
-// OPERATOR PRECEDENCE: 2
-const LogicalAnd = createToken({
-  name: 'LogicalAnd',
-  pattern: /&&/,
-  categories: BinaryOp,
-});
-
-// OPERATOR PRECEDENCE: 1
-const LogicalOr = createToken({
-  name: 'LogicalOr',
-  pattern: /\|\|/,
-  categories: BinaryOp,
-});
-
-export const Token = {
-  WhiteSpace,
-  Lookup,
-  NumberLiteral,
-  StringLiteral,
-  DblQuoteStringLiteral,
-  NestedDblQuoteStringLiteral,
-  SingleQuoteStringLiteral,
-  StringTemplateOpen,
-  StringTemplateClose,
-  InlineObjectExprOpen,
-  InlineObjectExprClose,
-  LParen,
-  RParen,
-  LCurly,
-  RCurly,
-  LSquare,
-  RSquare,
-  TernaryOper,
-  Dot,
-  Comma,
-  Colon,
-  True,
-  False,
-  Nullable,
-  Null,
-  Nil,
-  LogicalAnd,
-  MulOp,
-  AddOp,
-  RelOp,
-  LogicalOr,
-  BinaryOp,
-  ResolverIdentifier,
-  IdentifierName,
-};
-
-const objectExpressionMode = (close: TokenType) => {
-  return [
-    WhiteSpace,
-    Lookup,
-    NumberLiteral,
-    StringLiteral,
-    NestedDblQuoteStringLiteral,
-    SingleQuoteStringLiteral,
-    close,
-    LParen,
-    RParen,
-    LSquare,
-    RSquare,
-    TernaryOper,
-    Dot,
-    Comma,
-    Colon,
-    True,
-    False,
-    Nil,
-    Null,
-    Nullable,
-    LogicalAnd,
-    MulOp,
-    AddOp,
-    RelOp,
-    LogicalOr,
-    BinaryOp,
-    ResolverIdentifier,
-    IdentifierName,
-  ];
-};
-
-export const lexerDefinition: IMultiModeLexerDefinition = {
-  modes: {
-    default: [
-      WhiteSpace,
-      Lookup,
-      NumberLiteral,
-      StringTemplateOpen,
-      InlineObjectExprOpen,
-      StringLiteral,
-      DblQuoteStringLiteral,
-      SingleQuoteStringLiteral,
-      LParen,
-      RParen,
-      LCurly,
-      RCurly,
-      LSquare,
-      RSquare,
-      TernaryOper,
-      Dot,
-      Comma,
-      Colon,
-      True,
-      False,
-      Null,
-      Nil,
-      Nullable,
-      LogicalAnd,
-      MulOp,
-      AddOp,
-      RelOp,
-      LogicalOr,
-      BinaryOp,
-      ResolverIdentifier,
-      IdentifierName,
-    ],
-    string_template: objectExpressionMode(StringTemplateClose),
-    inline_object_expr: objectExpressionMode(InlineObjectExprClose),
-  },
-  defaultMode: 'default',
-};
-
-export const allTokens: TokenType[] = [...Object.values(Token)];
 
 /////////////////////////////
 //        PARSER           //
@@ -610,3 +227,14 @@ export class MappingParser extends Parser {
     this.CONSUME(Token.InlineObjectExprClose);
   });
 }
+
+// todo: get of rid of this temporary hack to fix issues for client production build
+// when angular cli's optimization options are enabled the process (which is using the library "Terser") removes the name of the classes
+// this causes that at runtime MappingParser.constructor.name has a value of an empty string ("") and Chevrotain depends on said constuctor.name
+// during runtime (https://github.com/SAP/chevrotain/blob/21b76452faf1df0413df83d7852f7b1d6534f7a4/packages/chevrotain/src/lang/lang_extensions.ts#L15)
+// Angular CLI doesn't expose by default a way to configure the optimizations it applies to the code nor we can define a `name` manually in the class
+// because NodeJS will complain because `name` is a readonly property automatically defined by Javascript.
+// Hence we're re-defining the class' `toString()` method so the fallback in Chevrotain's code can be used
+MappingParser.toString = function() {
+  return 'MappingParser';
+};
